Add surface type to PlayerSurface
