+++ pillarbox-ui/docs/README.md
@@ -34,7 +34,7 @@ fun SimplePlayer(player: Player){
 
 ### Create a simple player with controls and subtitles
 
-In this example we are drawing controls and subtitles on top of the player surface. To add subtitles use `ExoPlayerSubtitleView` and for controls
+In this example, we are drawing controls and subtitles on top of the player surface. To add subtitles use `ExoPlayerSubtitleView` and for controls
 you can use the Exoplayer version, `ExoPlayerControlView`.
 
 ```kotlin
@@ -52,7 +52,8 @@ fun MyPlayer(player: Player) {
             modifier = Modifier,
             player = player,
             scaleMode = ScaleMode.Fit,
-            defaultAspectRatio = defaultAspectRatio
+            defaultAspectRatio = defaultAspectRatio,
+            surfaceType = SurfaceType.Surface, // By default
         )
         ExoPlayerControlView(modifier = Modifier.matchParentSize(), player = player)
         ExoPlayerSubtitleView(modifier = Modifier.matchParentSize(), player = player)
@@ -67,6 +68,21 @@ In this example we use `ScaleMode.Fit` to fit the content to the parent containe
 - `ScaleMode.Fit` : Fit player content to the parent container and keep aspect ratio.
 - `ScaleMode.Fill` : Fill player content to the parent container.
 - `ScaleMode.Crop` : Crop player content inside the parent container and keep aspect ratio. Content outside the parent container will be clipped.
+
+## Surface types
+
+PlayerSurface allows choosing between multiple surface types:
+
+ - `SurfaceType.Surface` : The player is linked to a `SurfaceView` the default option.
+   This option is the more optimized version and allows to play 
+   any content including DRM protected content.
+ - `SurfaceType.Texture`: The player is linked to a `TextureView`.
+   This surface maybe interesting when dealing with animation and surface type has 
+   issue.
+ - `SurfaceType.Spherical` : The player is linked to a `SphericalGLSurfaceView`. This surface type is suited when playing 360° video content.
+
+The two last surface types are not suited when playing DRM protected content.
+
 
 ### Listen to player states
