+++ pillarbox-ui/src/main/java/ch/srgssr/pillarbox/ui/widget/player/PlayerSurface.kt
@@ -4,15 +4,8 @@
  */
 package ch.srgssr.pillarbox.ui.widget.player
 
-import android.content.Context
-import android.graphics.Canvas
-import android.os.Build
-import android.os.Handler
-import android.os.Looper
-import android.view.SurfaceControl
 import android.view.SurfaceView
-import android.window.SurfaceSyncGroup
-import androidx.annotation.RequiresApi
+import android.view.TextureView
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxScope
@@ -32,8 +25,8 @@ import androidx.compose.ui.draw.clipToBounds
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.drawscope.Stroke
-import androidx.compose.ui.viewinterop.AndroidView
 import androidx.media3.common.Player
+import androidx.media3.exoplayer.video.spherical.SphericalGLSurfaceView
 import ch.srgssr.pillarbox.player.tracks.videoTracks
 import ch.srgssr.pillarbox.ui.ScaleMode
 import ch.srgssr.pillarbox.ui.exoplayer.ExoPlayerSubtitleView
@@ -48,9 +41,11 @@ import ch.srgssr.pillarbox.ui.extension.getAspectRatioAsState
  * @param contentAlignment The "letterboxing" content alignment inside the parent. Only used when the aspect ratio is strictly positive.
  * @param defaultAspectRatio The aspect ratio to use while video is loading or for audio content.
  * @param displayDebugView When `true`, displays debug information on top of the surface. Only used when the aspect ratio is strictly positive.
+ * @param surfaceType Surface type to use. When playing DRM content, don't use [SurfaceType.Texture] or [SurfaceType.Spherical].
  * @param surfaceContent The Composable content to display on top of the [SurfaceView]. By default, render the subtitles. Only used when the aspect
  * ratio is strictly positive.
  */
+@Suppress("CyclomaticComplexMethod")
 @Composable
 fun PlayerSurface(
     player: Player,
@@ -59,6 +54,7 @@ fun PlayerSurface(
     contentAlignment: Alignment = Alignment.Center,
     defaultAspectRatio: Float? = null,
     displayDebugView: Boolean = false,
+    surfaceType: SurfaceType = SurfaceType.Surface,
     surfaceContent: @Composable (BoxScope.() -> Unit)? = { ExoPlayerSubtitleView(player = player) },
 ) {
     var lastKnownVideoAspectRatio by remember { mutableFloatStateOf(defaultAspectRatio ?: 1f) }
@@ -100,7 +96,11 @@ fun PlayerSurface(
             }
         }
 
-        AndroidPlayerSurfaceView(modifier = videoSurfaceModifier, player = player)
+        when (surfaceType) {
+            SurfaceType.Surface -> AndroidPlayerSurfaceView(modifier = videoSurfaceModifier, player = player)
+            SurfaceType.Texture -> AndroidPlayerTextureView(modifier = videoSurfaceModifier, player = player)
+            SurfaceType.Spherical -> AndroidSphericalSurfaceView(modifier = videoSurfaceModifier, player = player)
+        }
 
         surfaceContent?.let {
             val overlayModifier = if (scaleMode != ScaleMode.Crop) {
@@ -129,6 +129,26 @@ fun PlayerSurface(
 }
 
 /**
+ * Surface type
+ */
+enum class SurfaceType {
+    /**
+     * Render into a [SurfaceView].
+     */
+    Surface,
+
+    /**
+     * Render into a [TextureView], not compatible with DRM content.
+     */
+    Texture,
+
+    /**
+     * Render into a [SphericalGLSurfaceView] useful for 360° content.
+     */
+    Spherical,
+}
+
+/**
  * Debug player view
  *
  * @param modifier The modifier to use to layout.
@@ -152,110 +172,5 @@ private fun DebugPlayerView(modifier: Modifier) {
             color = Color.Magenta,
             style = Stroke(width = 4f),
         )
-    }
-}
-
-/**
- * Render the [player] content on a [SurfaceView].
- *
- * @param player The player to render on the SurfaceView.
- * @param modifier The modifier to be applied to the layout.
- */
-@Composable
-private fun AndroidPlayerSurfaceView(player: Player, modifier: Modifier = Modifier) {
-    AndroidView(
-        /*
-         * On some devices (Pixel 2 XL Android 11),
-         * the "black" background of the SurfaceView shows outside its bound.
-         */
-        modifier = modifier.clipToBounds(),
-        factory = { context ->
-            PlayerSurfaceView(context)
-        },
-        update = { view ->
-            view.player = player
-        },
-        onRelease = { view ->
-            view.player = null
-        },
-        onReset = { view ->
-            // onReset is called before `update` when the composable is reused with a different context.
-            view.player = null
-        }
-    )
-}
-
-/**
- * Player surface view
- */
-private class PlayerSurfaceView(context: Context) : SurfaceView(context) {
-    private val playerListener = PlayerListener()
-    private val surfaceSyncGroupV34 = when {
-        isInEditMode -> null
-        needSurfaceSyncWorkaround() -> SurfaceSyncGroupCompatV34()
-        else -> null
-    }
-
-    /**
-     * Player if null is passed just clear surface
-     */
-    var player: Player? = null
-        set(value) {
-            if (field != value) {
-                field?.clearVideoSurfaceView(this)
-                field?.removeListener(playerListener)
-                value?.setVideoSurfaceView(this)
-                value?.addListener(playerListener)
-            }
-            field = value
-        }
-
-    override fun dispatchDraw(canvas: Canvas) {
-        super.dispatchDraw(canvas)
-
-        if (needSurfaceSyncWorkaround()) {
-            surfaceSyncGroupV34?.maybeMarkSyncReadyAndClear()
-        }
-    }
-
-    // Workaround for a surface sync issue on API 34: https://github.com/androidx/media/issues/1237
-    // Imported from https://github.com/androidx/media/commit/30cb76269a67e09f6e1662ea9ead6aac70667028
-    private fun needSurfaceSyncWorkaround(): Boolean {
-        return Build.VERSION.SDK_INT == Build.VERSION_CODES.UPSIDE_DOWN_CAKE
-    }
-
-    private inner class PlayerListener : Player.Listener {
-        private val mainLooperHandler = Handler(Looper.getMainLooper())
-
-        override fun onSurfaceSizeChanged(width: Int, height: Int) {
-            if (needSurfaceSyncWorkaround()) {
-                surfaceSyncGroupV34?.postRegister(mainLooperHandler, this@PlayerSurfaceView)
-            }
-        }
-    }
-
-    @RequiresApi(Build.VERSION_CODES.UPSIDE_DOWN_CAKE)
-    private class SurfaceSyncGroupCompatV34 {
-        private var surfaceSyncGroup: SurfaceSyncGroup? = null
-
-        fun postRegister(
-            mainLooperHandler: Handler,
-            surfaceView: SurfaceView,
-        ) {
-            mainLooperHandler.post {
-                // The SurfaceView isn't attached to a window, so don't apply the workaround.
-                val rootSurfaceControl = surfaceView.getRootSurfaceControl() ?: return@post
-
-                surfaceSyncGroup = SurfaceSyncGroup("exo-sync-b-334901521")
-                surfaceSyncGroup?.add(rootSurfaceControl) {}
-                surfaceView.invalidate()
-                rootSurfaceControl.applyTransactionOnDraw(SurfaceControl.Transaction())
-            }
-        }
-
-        fun maybeMarkSyncReadyAndClear() {
-            surfaceSyncGroup?.markSyncReady()
-            surfaceSyncGroup = null
-        }
     }
 }
